code/backend/requirements.txt
Flask
Flask-RESTful
opencv-python
pytest-shutil
Pillow
keras==2.8
tensorflow==2.8.0
mariadb
matplotlib
pandas
scikit-learn
scipy
virtualenv
numpy
tqdm 
